Change seeder (insert -> create)

+++ database/migrations/2023_06_01_234455_create_meetings_table.php
@@ -21,7 +21,7 @@ class CreateMeetingsTable extends Migration
             $table->unsignedBigInteger('level_id');
             $table->string('title');
             $table->timestamp('date');
-            $table->unsignedBigInteger('start_at');
+            $table->integer('start_at');
             $table->integer('status_id')->default(1); //1:stand by
 
             $table->softDeletes();
